# docs/openapi.yaml
openapi: 3.0.3
info:
  title: A2A API Gateway
  description: |
    نظام تكامل شامل للتطبيقات (Application-to-Application) يربط بين التطبيقات المحلية ونظام الأوتوبيلوت وبوت تيليجرام.
    
    ## الميزات الرئيسية
    - API Gateway موحد لإدارة جميع الاتصالات
    - Message Broker للتواصل غير المتزامن
    - تكامل الأوتوبيلوت لإدارة المهام التلقائية
    - تكامل تيليجرام للتفاعل مع المستخدمين
    - نظام مراقبة شامل مع Prometheus و Grafana
    - أمان متقدم مع JWT و Rate Limiting
  version: 1.0.0
  contact:
    name: AuraOS Team
    email: support@auraos.com
    url: https://auraos.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.auraos.com
    description: Production server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/messages/publish:
    post:
      summary: Publish message
      description: Publish a message to a topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - topic
                - type
                - target
                - payload
              properties:
                topic:
                  type: string
                  example: telegram.updates
                type:
                  type: string
                  example: notification
                target:
                  type: string
                  example: telegram
                payload:
                  type: object
                  example:
                    message: Hello from A2A System!
                    chatId: 123456789
                priority:
                  type: string
                  enum: [low, normal, high, critical]
                  default: normal
                correlationId:
                  type: string
                  example: req_123456
      responses:
        '200':
          description: Message published successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  messageId:
                    type: string
                    example: msg_1234567890_abc123
                  topic:
                    type: string
                    example: telegram.updates
                  timestamp:
                    type: string
                    format: date-time
                  message:
                    type: string
                    example: Message published successfully
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/messages/subscribe:
    post:
      summary: Subscribe to topic
      description: Subscribe to a topic for receiving messages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - topic
              properties:
                topic:
                  type: string
                  example: autopilot.events
                handler:
                  type: string
                  format: uri
                  example: https://your-app.com/webhook
      responses:
        '200':
          description: Successfully subscribed to topic
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  topic:
                    type: string
                    example: autopilot.events
                  handler:
                    type: string
                    example: https://your-app.com/webhook
                  message:
                    type: string
                    example: Successfully subscribed to topic
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/messages/subscribe/{topic}:
    delete:
      summary: Unsubscribe from topic
      description: Unsubscribe from a topic
      parameters:
        - name: topic
          in: path
          required: true
          schema:
            type: string
          example: autopilot.events
      responses:
        '200':
          description: Successfully unsubscribed from topic
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  topic:
                    type: string
                    example: autopilot.events
                  message:
                    type: string
                    example: Successfully unsubscribed from topic
        '404':
          description: Topic not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: API key for service authentication
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Validation failed
        message:
          type: string
          example: Invalid request parameters
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: username
              message:
                type: string
                example: Username is required
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          example: req_1234567890_abc123
...